tests/wakelock/wakelock-api.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WakeLock api Test</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/wake-lock/">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

promise_test(t => {
  assert_true("getWakeLock" in navigator);
  return navigator.getWakeLock("screen")
  .then(wakeLock => {
    assert_true(wakeLock instanceof WakeLock, "wakeLock is a WakeLock");
    assert_true(wakeLock.type instanceof WakeLockType, "wakeLock.type is a WakeLockType");
    assert_equals(typeof wakeLock.active, "boolean", "the type of wakeLock.active is a boolean");
  });
}, "Test that the WakeLock API is correct");

</script>
